requirements.txt
fastapi
uvicorn
pydantic
openai
supabase
python-dotenv